moved api views to api.py
