+++ parkstay/urls.py
@@ -1,26 +1,24 @@
 from django.conf.urls import url, include
 from rest_framework import routers
-from parkstay import views
+from parkstay import views, api
 from parkstay.admin import admin
 
 from ledger.urls import urlpatterns as ledger_patterns
 
 # API patterns
 router = routers.DefaultRouter()
-router.register(r'campgrounds', views.CampgroundViewSet)
+router.register(r'campgrounds', api.CampgroundViewSet)
 #router.register(r'campsites', views.CampsiteViewSet)
-router.register(r'campsite_bookings', views.CampsiteBookingViewSet)
-router.register(r'promo_areas',views.PromoAreaViewSet)
-router.register(r'parks',views.ParkViewSet)
-router.register(r'campground_feature',views.FeatureViewSet)
-router.register(r'regions',views.RegionViewSet)
+router.register(r'campsite_bookings', api.CampsiteBookingViewSet)
+router.register(r'promo_areas',api.PromoAreaViewSet)
+router.register(r'parks',api.ParkViewSet)
+router.register(r'campground_feature',api.FeatureViewSet)
+router.register(r'regions',api.RegionViewSet)
 #router.register(r'campsite_classes',views.CampsiteClassViewSet)
-router.register(r'booking',views.BookingViewSet)
-router.register(r'campsite_rate',views.CampsiteRateViewSet)
+router.register(r'booking',api.BookingViewSet)
+router.register(r'campsite_rate',api.CampsiteRateViewSet)
 
 api_patterns = [
-#    url(r'^api/campsites/$', views.get_campsite_bookings, name='get_campsite_bookings'),
-#    url(r'^api/campsite_classes/$', views.get_campsite_class_bookings, name='get_campsite_class_bookings'),
     url(r'api/',include(router.urls))
 ]
